Split out initial files
